requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-modelcluster==5.0.2
django-money==1.0
django-multiselectfield==0.1.12
django-pagedown==2.1.2
django-resized==0.3.11
django-rest-knox==4.1.0
django-summernote==0.8.11.6
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
draftjs-exporter==2.1.7
field==0.2.0
gunicorn==20.0.4
heroku==0.1.4
html5lib==1.0.1
idna==2.9
l18n==2018.5
lxml==4.5.0
markdown2==2.3.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.4
py-moneyed==0.8.0
pycparser==2.20
pygame==1.9.6
PyJWT==1.7.1
python-dateutil==1.5
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.48.0
Unidecode==1.1.1
urllib3==1.25.8
wagtail==2.9
webencodings==0.5.1
whitenoise==5.0.1
Willow==1.3
XlsxWriter==1.2.9
